Move cobra tests to top level suite
