+++ cobra-hello/cobra_hello_suite_test.go
@@ -1,0 +1,13 @@
+package main_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestCobraHello(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "main")
+}
